+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/feature/messages/list/adapter/viewholder/attachment/AttachmentFactoryManager.kt
@@ -19,6 +19,7 @@ package io.getstream.chat.android.ui.feature.messages.list.adapter.viewholder.at
 import android.view.ViewGroup
 import io.getstream.chat.android.models.Message
 import io.getstream.chat.android.ui.feature.messages.list.adapter.MessageListListenerContainer
+import io.getstream.chat.android.ui.feature.messages.list.adapter.MessageListListeners
 
 /**
  * A manager for registered custom attachment factories.
@@ -47,9 +48,23 @@ public class AttachmentFactoryManager(
      * @param parent The parent View where the attachment content view is supposed to be placed.
      * @return An inner ViewHolder with the attachment content view.
      */
+    @Deprecated(
+        message = "Use createViewHolder(message: Message, listeners: MessageListListeners?, parent: ViewGroup) instead",
+        replaceWith = ReplaceWith("createViewHolder(message, listeners, parent)"),
+        level = DeprecationLevel.WARNING,
+    )
     public fun createViewHolder(
         message: Message,
         listeners: MessageListListenerContainer?,
+        parent: ViewGroup,
+    ): InnerAttachmentViewHolder {
+        val factory = attachmentFactories.first { it.canHandle(message) }
+        return factory.createViewHolder(message, listeners, parent)
+    }
+
+    public fun createViewHolder(
+        message: Message,
+        listeners: MessageListListeners?,
         parent: ViewGroup,
     ): InnerAttachmentViewHolder {
         val factory = attachmentFactories.first { it.canHandle(message) }